third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-panel-restore-selection-when-node-comes-later-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify that last selected element is restored properly later, even if it failed to do so once.


Running: selectNode
Selected node: 'span'

Running: firstReloadWithoutNodeInDOM
Page reloaded.
Selected node: 'body'

Running: secondReloadWithNodeInDOM
Page reloaded.
Selected node: 'span'
